Fix: Parsing nested code blocks off-by-one
After testing the new Graphviz debugging on Examples seq100, it was
clear it could not parse any of the loop nodes and function calls
correctly. This issue should now be fixed. Hopefully for other code
blocks too.

If other code blocks are messed up. Look at the implementation of
`return` and `loop`. These are now both correctly implemented
`return` being a simple expression, and `loop` being a complex inner block

Signed-off-by: Lilith <48443491+Stroopwafe1@users.noreply.github.com>