Feat/Fix: Struct, Class, and Function parsing working fully!!
Needed to fix function parsing with multiple arguments, that was
apparently broken and I hadn't tested that.
Structs are now fully working properly. I had to change the code for
the field list slightly as multiple property names was bugged with
off-by-one errors.
Same goes for the class, but the class had extra problems due to
having both fields and functions

Signed-off-by: Lilith <48443491+Stroopwafe1@users.noreply.github.com>